requirements.txt
openai>=1.30.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
rich>=13.7.0
click>=8.1.7
pydantic>=2.7.0
pydantic-settings>=2.3.0
